update demo; remove unuse code;
