+++ Sources/HandVector/HandVectorTool.swift
@@ -46,26 +46,31 @@ public class HandVectorTool {
         }
     }
     @MainActor
+    public func addHand(from handAnchor: HandAnchor, to: Entity?, filter: CollisionFilter = .default) async {
+        if let handInfo = HVHandInfo(handAnchor: handAnchor) {
+            if handInfo.chirality == .left {
+                leftHandVector = handInfo
+                let left = generateHandEntity(from: handInfo, filter: filter)
+                to?.addChild(left)
+                self.left = left
+            } else if handInfo.chirality == .right {
+                rightHandVector = handInfo
+                let right = generateHandEntity(from: handInfo, filter: filter)
+                to?.addChild(right)
+                self.right = right
+            }
+        }
+    }
+    @MainActor
     public func updateHand(from handAnchor: HandAnchor, filter: CollisionFilter = .default) async {
         if let handInfo = HVHandInfo(handAnchor: handAnchor) {
             if handInfo.chirality == .left {
                 leftHandVector = handInfo
-                if left == nil {
-                    left = generateHandEntity(from: handInfo, filter: filter)
-                } else {
-                    updateHandEntity(from: handInfo, to: left!)
-                }
-//                let shape = handInfo.calculateFingerShape(finger: .thump)
+                updateHandEntity(from: handInfo, to: left!)
             } else if handInfo.chirality == .right { // Update right hand info.
                 rightHandVector = handInfo
-                if right == nil {
-                    right = generateHandEntity(from: handInfo, filter: filter)
-                } else {
-                    updateHandEntity(from: handInfo, to: right!)
-                }
-                
+                updateHandEntity(from: handInfo, to: right!)
             }
-            
         }
     }
     public func removeHand(from handAnchor: HandAnchor) {
@@ -80,45 +85,8 @@ public class HandVectorTool {
         }
     }
     
-//    private func generateHandEntity(from handAnchor: HandAnchor, filter: CollisionFilter = .default) -> Entity {
-//        let hand = Entity()
-//        hand.name = handAnchor.chirality == .left ? "leftHand" : "rightHand"
-//        hand.transform.matrix = handAnchor.originFromAnchorTransform
-//        
-//        if let skeleton = handAnchor.handSkeleton {
-//            let wm = SimpleMaterial(color: .white, isMetallic: false)
-//            let rm = SimpleMaterial(color: .red, isMetallic: false)
-//            for joint in skeleton.allJoints {
-//                let m = (joint.name == .wrist || joint.name == .forearmWrist) ? rm : wm
-//                let modelEntity = ModelEntity(mesh: .generateSphere(radius: 0.01), materials: [m])
-//                modelEntity.transform.matrix = joint.anchorFromJointTransform
-//                modelEntity.name = joint.name.codableName.rawValue + "-model"
-//                modelEntity.isEnabled = isSkeletonVisible
-//                hand.addChild(modelEntity)
-//                
-//                let collisionEntity = Entity()
-//                collisionEntity.components.set(CollisionComponent(shapes: [.generateSphere(radius: 0.01)], filter: filter))
-//                collisionEntity.transform.matrix = joint.anchorFromJointTransform
-//                collisionEntity.name = joint.name.codableName.rawValue + "-collision"
-//                hand.addChild(collisionEntity)
-//            }
-//        }
-//        return hand
-//    }
-//    private func updateHandEntity(from handAnchor: HandAnchor, to: Entity) {
-//        to.transform.matrix = handAnchor.originFromAnchorTransform
-//        if let skeleton = handAnchor.handSkeleton {
-//            for joint in skeleton.allJoints {
-//                let modelEntity = to.findEntity(named: joint.name.codableName.rawValue + "-model")
-//                modelEntity?.transform.matrix = joint.anchorFromJointTransform
-//                modelEntity?.isEnabled = isSkeletonVisible
-//                
-//                let collisionEntity = to.findEntity(named: joint.name.codableName.rawValue + "-collision")
-//                collisionEntity?.transform.matrix = joint.anchorFromJointTransform
-//            }
-//        }
-//    }
-    
+
+    public static let simHandPositionY: Float = 1.4
     @MainActor
     public func updateHand(from simHand: SimHand, filter: CollisionFilter = .default) async {
         let handVectors = simHand.convertToHandVector(offset: .init(0, Self.simHandPositionY, -0.2))
@@ -189,8 +157,6 @@ public class HandVectorTool {
         }
         
     }
-    
-    public static let simHandPositionY: Float = 1.4
 }
